requirements.txt
pandas==1.5.1
Pillow==9.4.0
plotly==5.13.0
prophet==1.1.2
streamlit==1.17.0
yfinance==0.1.87
